fixes for label and correctness count display
